I believe that pfc_n_cst_measureconversion in pfcapsrv.pbl should be marked to be autoinstatiated. This object is used in pfc_n_cst_pbunitconversion. But nowhere in this object pfc_n_cst_measureconversion (invo_measurementConversion) is created or destroyed. In many accasions pfc_n_cst_pbunitconversion calls functions declared in pfc_n_cst_measureconversion. Actually pfc_n_cst_measureconversion doesn't have anything except from object functions doing unit conversions.
In our case, we had an issue with sort service in datawindow (based on column header). In latest pfcs it puts an arrow in the column used to sort data, showing if it's an ascending or descending sort. In some cases some functions are used to calculate the position to place that symbol. As the object is not instantiated the application will crash at this point,